Renamed telecommand handling into telecommunication
