+++ libs/CMakeLists.txt
@@ -13,7 +13,7 @@ add_subdirectory(adcs)
 add_subdirectory(power)
 add_subdirectory(terminal)
 add_subdirectory(assert)
-add_subdirectory(telecommand_handling)
+add_subdirectory(telecommunication)
 add_subdirectory(efm_support)
 add_subdirectory(obc)
 add_subdirectory(settings)
