+++ libs/obc/communication/CMakeLists.txt
@@ -9,7 +9,7 @@ add_library(${NAME} STATIC ${SOURCES})
 target_link_libraries(${NAME} 
 	base
 	comm
-	telecommand_handling
+	telecommunication
 	settings
 )
