+++ libs/telecommunication/include/telecommunication/telecommand_handling.h
@@ -1,5 +1,5 @@
-#ifndef LIBS_TELECOMMAND_HANDLING_INCLUDE_TELECOMMAND_HANDLING_TELECOMMAND_HANDLING_H_
-#define LIBS_TELECOMMAND_HANDLING_INCLUDE_TELECOMMAND_HANDLING_TELECOMMAND_HANDLING_H_
+#ifndef LIBS_TELECOMMUNICATION_INCLUDE_TELECOMMUNICATION_TELECOMMAND_HANDLING_H_
+#define LIBS_TELECOMMUNICATION_INCLUDE_TELECOMMUNICATION_TELECOMMAND_HANDLING_H_
 
 #include <cstdint>
 #include <gsl/span>
@@ -142,4 +142,4 @@ namespace telecommands
 }
 /** @} */
 
-#endif /* LIBS_TELECOMMAND_HANDLING_INCLUDE_TELECOMMAND_HANDLING_TELECOMMAND_HANDLING_H_ */
+#endif /* LIBS_TELECOMMUNICATION_INCLUDE_TELECOMMUNICATION_TELECOMMAND_HANDLING_H_ */
