+++ libs/telecommunication/include/telecommunication/uplink.h
@@ -1,9 +1,8 @@
-#ifndef LIBS_TELECOMMAND_HANDLING_INCLUDE_TELECOMMAND_HANDLING_UPLINK_H_
-#define LIBS_TELECOMMAND_HANDLING_INCLUDE_TELECOMMAND_HANDLING_UPLINK_H_
+#ifndef LIBS_TELECOMMUNICATION_INCLUDE_TELECOMMUNICATION_UPLINK_H_
+#define LIBS_TELECOMMUNICATION_INCLUDE_TELECOMMUNICATION_UPLINK_H_
 
 #include <cstdint>
 #include <gsl/span>
-
 #include "telecommand_handling.h"
 
 namespace telecommands
@@ -35,4 +34,4 @@ namespace telecommands
     };
 }
 
-#endif /* LIBS_TELECOMMAND_HANDLING_INCLUDE_TELECOMMAND_HANDLING_UPLINK_H_ */
+#endif /* LIBS_TELECOMMUNICATION_INCLUDE_TELECOMMUNICATION_UPLINK_H_ */
