+++ libs/telecommunication/uplink.cpp
@@ -1,7 +1,6 @@
+#include "uplink.h"
 #include <algorithm>
 #include "base/reader.h"
-
-#include "uplink.h"
 
 using std::uint8_t;
 using std::copy;
