+++ unit_tests/CMakeLists.txt
@@ -84,7 +84,7 @@ target_link_libraries(${NAME}
     yaffs
     storage
     comm
-    telecommand_handling
+    telecommunication
     fs    
     base
     mission
